Mv'ed parsing and formatting logic out of `tracks/hic/hic.straw.ts` to new `parse.genome.ts`
