+++ client/tracks/hic/test/hic.integration.spec.ts
@@ -2,7 +2,8 @@ import tape from 'tape'
 import * as d3s from 'd3-selection'
 import { dofetch2 } from '../../../common/dofetch.js'
 import { hicData } from './hicData.ts'
-import { hicparsefile, hicparsestat } from '../../../tracks/hic/hic.straw.ts'
+import { hicparsefile } from '../../../tracks/hic/hic.straw.ts'
+import { hicparsestat } from '../../../tracks/hic/parse.genome.ts'
 // import { runproteinpaint } from '../../../test/front.helpers.js'
 // import { HicRunProteinPaintTrackArgs } from '../../../types/hic.ts'
